// src/generators/content-types.ts
interface ContentTypeGenData {
  type: string;
  hasParagraphs: boolean;
  hasMediaLibrary: boolean;
  hasCkeditor: boolean;
  requiredNodeFields: ResolvedFieldInfo[];
  paragraphFields: ParagraphFieldConfig[];
  /** Primary paragraph field (most allowed types). */
  primaryParagraphField?: ParagraphFieldConfig;
  /** Secondary paragraph fields. */
  secondaryParagraphFields: ParagraphFieldConfig[];
}